+++ containers/feeds/feeds.module.css
@@ -1,16 +1,18 @@
 .usersSection {
-    height: 1000px;
-    width: 80%;
+    width: 60%;
     margin: auto;
     display: flex;
     flex-flow: row;
 }
 
-
 .feeds {
     width: 80%;
     margin-top: 3%;
     margin: auto;
+}
+
+._hr {
+    opacity: 0.3;
 }
 
 .count {
@@ -27,10 +29,49 @@
 .profileItems {
     display: flex;
     flex-flow: row;
-    height: 130px;
-    /* margin-bottom: 5%; */
+    height: 100px;
+    margin-top: 3%;
+    margin-bottom: 3%;
 }
 .name_ {
     margin-left: 5%;
     color: #58A6FF;
+}
+.followBtn {
+    margin-left: 15%;
+    height: 30px;
+    width: 20%;
+    margin-top: 2%;
+    background: #FAFBFC;
+    border: 1px solid #D9DBDB;
+    outline: none;
+    border-radius: 5px;
+    cursor: pointer;
+}
+.stack {
+    display: flex;
+    flex-flow: row;
+    width: 50%;
+}
+.btnDivDown {
+    width: 60%;
+    margin: auto;
+}
+.btnDiv {
+    margin-top: 3%;
+    text-align: center;
+}
+.a_ {
+    color: #24292E;
+}
+.nav {
+    width: 10%;
+    height: 30px;
+    margin-left: 5%;
+    margin-bottom: 10%;
+    background: #FAFBFC;
+    border: 1px solid #D9DBDB;
+    outline: none;
+    border-radius: 5px;
+    cursor: pointer;
 }
